change class detailcoffee
